+++ android/src/main/java/com/example/sms_mms/SmsMmsPlugin.java
@@ -1,27 +1,12 @@
 package com.example.sms_mms;
 
-import static android.content.Intent.CATEGORY_APP_MESSAGING;
-import static android.content.Intent.CATEGORY_DEFAULT;
+import static android.content.Intent.EXTRA_STREAM;
 import static android.content.Intent.FLAG_ACTIVITY_NEW_TASK;
 
-import android.app.Activity;
-import android.content.ActivityNotFoundException;
-import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
-import android.content.pm.ActivityInfo;
-import android.content.pm.PackageManager;
-import android.content.pm.ResolveInfo;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
 import android.net.Uri;
-import android.os.Build;
-import android.os.Bundle;
-import android.os.CountDownTimer;
-import android.os.Environment;
 import android.provider.Telephony;
-import android.telephony.SmsManager;
-import android.util.Base64;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
@@ -29,19 +14,9 @@ import androidx.annotation.Nullable;
 import androidx.core.content.FileProvider;
 
 import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.net.MalformedURLException;
-import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
-import java.util.Objects;
 
-import io.flutter.Log;
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
-import io.flutter.embedding.engine.plugins.activity.ActivityAware;
-import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
@@ -51,20 +26,17 @@ import io.flutter.plugin.common.MethodChannel.Result;
  * SmsMmsPlugin
  */
 public class SmsMmsPlugin implements FlutterPlugin, MethodCallHandler {
-    /// The MethodChannel that will the communication between Flutter and native Android
-    ///
-    /// This local reference serves to register the plugin with the Flutter Engine and unregister it
-    /// when the Flutter Engine is detached from the Activity
-    private MethodChannel channel;
-
-    String DEFAULT_MESSAGE_PACKAGE_NAME = "com.google.android.apps.messaging";
 
     @Nullable
     private Context context;
 
     @Override
     public void onAttachedToEngine(@NonNull FlutterPluginBinding flutterPluginBinding) {
-        channel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(), "sms_mms");
+        /// The MethodChannel that will the communication between Flutter and native Android
+        ///
+        /// This local reference serves to register the plugin with the Flutter Engine and unregister it
+        /// when the Flutter Engine is detached from the Activity
+        MethodChannel channel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(), "sms_mms");
         context = flutterPluginBinding.getApplicationContext();
         channel.setMethodCallHandler(this);
     }
@@ -88,16 +60,32 @@ public class SmsMmsPlugin implements FlutterPlugin, MethodCallHandler {
                 addressString.append(address.get(i)).append(";");
             }
 
-            sendSMS(filePath, addressString.toString(), message);
+            try {
+                sendSMS(filePath, addressString.toString(), message);
+            } catch (Exception e) {
+                Toast.makeText(context, e.getMessage(), Toast.LENGTH_SHORT).show();
+            }
 
         } else {
             result.notImplemented();
         }
     }
 
-    protected void sendSMS(@Nullable String filePath, String address, String message) {
+    protected void sendSMS(@Nullable String filePath, String address, String message) throws Exception {
 
         @Nullable Uri fileUri = null;
+        String DEFAULT_MESSAGE_PACKAGE_NAME = "";
+
+
+        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.KITKAT) {
+            String packageName = Telephony.Sms.getDefaultSmsPackage(context);
+            if(packageName != null) {
+                DEFAULT_MESSAGE_PACKAGE_NAME = packageName;
+            }
+            if(packageName == null) {
+                throw new Exception("Default sms app not found");
+            }
+        }
 
         if (filePath != null) {
             File file = new File(filePath);
@@ -106,26 +94,40 @@ public class SmsMmsPlugin implements FlutterPlugin, MethodCallHandler {
             fileUri = FileProvider.getUriForFile(context, providerAuthority, file);
         }
 
-        Intent smsIntent = new Intent(Intent.ACTION_VIEW);
+        String action = filePath != null ? Intent.ACTION_SEND : Intent.ACTION_VIEW;
 
-        smsIntent.addFlags(FLAG_ACTIVITY_NEW_TASK);
-        smsIntent.setData(Uri.parse("mmsto:"));
-        smsIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
-        smsIntent.addFlags(Intent.FLAG_GRANT_WRITE_URI_PERMISSION);
-        smsIntent.addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY);
-        smsIntent.putExtra("address", address);
-        smsIntent.putExtra("sms_body", message);
 
-        try {
-            if (filePath != null) {
-                smsIntent.putExtra(Intent.EXTRA_STREAM, fileUri);
+        Intent shareIntent = new Intent(action);
+        shareIntent.addFlags(FLAG_ACTIVITY_NEW_TASK);
+        shareIntent.putExtra("sms_body", message);
+
+        if(filePath != null) {
+            shareIntent.setData(fileUri);
+            Uri returnUri = shareIntent.getData();
+            String mimeType = context.getContentResolver().getType(returnUri);
+            shareIntent.setType(mimeType);
+            shareIntent.putExtra(EXTRA_STREAM, fileUri);
+            shareIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+            if (context != null) {
                 context.grantUriPermission(DEFAULT_MESSAGE_PACKAGE_NAME, fileUri, Intent.FLAG_GRANT_WRITE_URI_PERMISSION | Intent.FLAG_GRANT_READ_URI_PERMISSION);
             }
-            context.startActivity(smsIntent);
-        } catch (android.content.ActivityNotFoundException ex) {
-            Log.e("ERROR", Objects.requireNonNull(ex.getMessage()));
-            Toast.makeText(context,
-                    "SMS failed, please try again later.", Toast.LENGTH_SHORT).show();
+            shareIntent.putExtra("address", address);
+        } else {
+            shareIntent.setData(Uri.fromParts("sms", address, null));
+        }
+
+        shareIntent.setPackage(DEFAULT_MESSAGE_PACKAGE_NAME);
+
+        try {
+            if (context != null) {
+                context.startActivity(shareIntent);
+            }
+        } catch (Exception e) {
+            if(filePath != null) {
+                throw new Exception("File format not supported by Messages");
+            } else {
+                throw new Exception("Unable to launch Messages");
+            }
         }
     }
 }